This follows #39436

I tried building it on an ARM machine. Make errors, presumably when compiling SparseSuite:
```
Base  ───────────── 40.682409 seconds
ArgTools  ─────────  9.163746 seconds
Artifacts  ────────  0.128767 seconds
Base64  ───────────  0.192947 seconds
CRC32c  ───────────  0.009570 seconds
FileWatching  ─────  0.134998 seconds
Libdl  ────────────  0.002042 seconds
Logging  ──────────  0.043996 seconds
Mmap  ─────────────  0.109363 seconds
NetworkOptions  ───  0.085440 seconds
SHA  ──────────────  0.283725 seconds
Serialization  ────  0.462806 seconds
Sockets  ──────────  0.431932 seconds
Unicode  ──────────  0.008099 seconds
DelimitedFiles  ───  0.138467 seconds
LinearAlgebra  ──── 10.231614 seconds
Markdown  ─────────  1.501900 seconds
Printf  ───────────  0.442666 seconds
Random  ───────────  0.679215 seconds
Tar  ──────────────  0.385961 seconds
Dates  ────────────  2.891745 seconds
Distributed  ──────  1.154808 seconds
Future  ───────────  0.006687 seconds
InteractiveUtils  ─  0.918014 seconds
LibGit2  ──────────  1.798731 seconds
Profile  ──────────  0.377027 seconds
SparseArrays  ─────  3.905063 seconds
UUIDs  ────────────  0.019885 seconds
REPL  ─────────────  6.763271 seconds
SharedArrays  ─────  0.189820 seconds
Statistics  ───────  0.232903 seconds
error during bootstrap:
LoadError("sysimg.jl", 19, LoadError("/home/mbesanco/src_julia/usr/share/julia/stdlib/v1.7/SuiteSparse/src/SuiteSparse.jl", 3, LoadError("/home/mbesanco/src_julia/usr/share/julia/stdlib/v1.7/SuiteSparse/src/umfpack.jl", 3, ErrorException("could not load library \"libsuitesparse_wrapper\"\nlibopenblas.so: cannot open shared object file : No such file or directory"))))
jl_errorf at /home/mbesanco/src_julia/src/rtutils.c:77
jl_load_dynamic_library at /home/mbesanco/src_julia/src/dlload.c:258
jl_get_library_ at /home/mbesanco/src_julia/src/runtime_ccall.cpp:52
jl_get_library_ at /home/mbesanco/src_julia/src/runtime_ccall.cpp:33 [inlined]
jl_load_and_lookup at /home/mbesanco/src_julia/src/runtime_ccall.cpp:63
unknown function (ip: 0xffff7c7d81fb)
top-level scope at /home/mbesanco/src_julia/usr/share/julia/stdlib/v1.7/SuiteSparse/src/umfpack.jl:67
jl_toplevel_eval_flex at /home/mbesanco/src_julia/src/toplevel.c:875
jl_eval_module_expr at /home/mbesanco/src_julia/src/toplevel.c:195 [inlined]
jl_toplevel_eval_flex at /home/mbesanco/src_julia/src/toplevel.c:672
jl_toplevel_eval_flex at /home/mbesanco/src_julia/src/toplevel.c:829
jl_toplevel_eval_in at /home/mbesanco/src_julia/src/toplevel.c:936
eval at ./boot.jl:369 [inlined]
include_string at ./loading.jl:1094
_jl_invoke at /home/mbesanco/src_julia/src/gf.c:2242 [inlined]
jl_apply_generic at /home/mbesanco/src_julia/src/gf.c:2424
_include at ./loading.jl:1148
include at ./Base.jl:386
_jl_invoke at /home/mbesanco/src_julia/src/gf.c:2242 [inlined]
jl_apply_generic at /home/mbesanco/src_julia/src/gf.c:2424
jl_apply at /home/mbesanco/src_julia/src/julia.h:1738 [inlined]
jl_f__call_latest at /home/mbesanco/src_julia/src/builtins.c:732
include at /home/mbesanco/src_julia/usr/share/julia/stdlib/v1.7/SuiteSparse/src/SuiteSparse.jl:3
_jl_invoke at /home/mbesanco/src_julia/src/gf.c:2242 [inlined]
jl_apply_generic at /home/mbesanco/src_julia/src/gf.c:2424
jl_apply at /home/mbesanco/src_julia/src/julia.h:1738 [inlined]
do_call at /home/mbesanco/src_julia/src/interpreter.c:117
eval_value at /home/mbesanco/src_julia/src/interpreter.c:206
eval_stmt_value at /home/mbesanco/src_julia/src/interpreter.c:157 [inlined]
eval_body at /home/mbesanco/src_julia/src/interpreter.c:588
jl_interpret_toplevel_thunk at /home/mbesanco/src_julia/src/interpreter.c:717
top-level scope at /home/mbesanco/src_julia/usr/share/julia/stdlib/v1.7/SuiteSparse/src/SuiteSparse.jl:25
jl_toplevel_eval_flex at /home/mbesanco/src_julia/src/toplevel.c:884
jl_eval_module_expr at /home/mbesanco/src_julia/src/toplevel.c:195 [inlined]
jl_toplevel_eval_flex at /home/mbesanco/src_julia/src/toplevel.c:672
jl_toplevel_eval_flex at /home/mbesanco/src_julia/src/toplevel.c:829
jl_toplevel_eval_in at /home/mbesanco/src_julia/src/toplevel.c:936
eval at ./boot.jl:369 [inlined]
include_string at ./loading.jl:1094
_jl_invoke at /home/mbesanco/src_julia/src/gf.c:2242 [inlined]
jl_apply_generic at /home/mbesanco/src_julia/src/gf.c:2424
_include at ./loading.jl:1148
include at ./Base.jl:386 [inlined]
_require at ./loading.jl:1050
require at ./loading.jl:914
require at ./loading.jl:901
_jl_invoke at /home/mbesanco/src_julia/src/gf.c:2242 [inlined]
jl_apply_generic at /home/mbesanco/src_julia/src/gf.c:2424
macro expansion at ./timing.jl:279 [inlined]
macro expansion at ./sysimg.jl:90 [inlined]
macro expansion at ./timing.jl:279 [inlined]
top-level scope at ./sysimg.jl:89
jl_toplevel_eval_flex at /home/mbesanco/src_julia/src/toplevel.c:875
jl_parse_eval_all at /home/mbesanco/src_julia/src/toplevel.c:1010
jl_load_ at /home/mbesanco/src_julia/src/toplevel.c:1057
jl_load at /home/mbesanco/src_julia/src/toplevel.c:1070
exec_program at /home/mbesanco/src_julia/src/jlapi.c:500
true_main at /home/mbesanco/src_julia/src/jlapi.c:572
repl_entrypoint at /home/mbesanco/src_julia/src/jlapi.c:702
main at /home/mbesanco/src_julia/cli/loader_exe.c:51
__libc_start_main at /lib/aarch64-linux-gnu/libc.so.6 (unknown line)
_start at /home/mbesanco/src_julia/usr/bin/julia (unknown line)
```