@matbesancon 2021-02-17
Correction, this is not when building LinearAlgebra, which is above:
```
Base  ───────────── 40.682409 seconds
ArgTools  ─────────  9.163746 seconds
Artifacts  ────────  0.128767 seconds
Base64  ───────────  0.192947 seconds
CRC32c  ───────────  0.009570 seconds
FileWatching  ─────  0.134998 seconds
Libdl  ────────────  0.002042 seconds
Logging  ──────────  0.043996 seconds
Mmap  ─────────────  0.109363 seconds
NetworkOptions  ───  0.085440 seconds
SHA  ──────────────  0.283725 seconds
Serialization  ────  0.462806 seconds
Sockets  ──────────  0.431932 seconds
Unicode  ──────────  0.008099 seconds
DelimitedFiles  ───  0.138467 seconds
LinearAlgebra  ──── 10.231614 seconds
Markdown  ─────────  1.501900 seconds
Printf  ───────────  0.442666 seconds
Random  ───────────  0.679215 seconds
Tar  ──────────────  0.385961 seconds
Dates  ────────────  2.891745 seconds
Distributed  ──────  1.154808 seconds
Future  ───────────  0.006687 seconds
InteractiveUtils  ─  0.918014 seconds
LibGit2  ──────────  1.798731 seconds
Profile  ──────────  0.377027 seconds
SparseArrays  ─────  3.905063 seconds
UUIDs  ────────────  0.019885 seconds
REPL  ─────────────  6.763271 seconds
SharedArrays  ─────  0.189820 seconds
Statistics  ───────  0.232903 seconds
```