@ViralBShah 2021-02-18
It's because SuiteSparse is linked against 32-bit openblas currently.